+++ base/Module.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * @link    http://hiqdev.com/hipanel
+ * @license http://hiqdev.com/hipanel/license
+ * @copyright Copyright (c) 2015 HiQDev
+ */
+
+namespace hipanel\base;
+
+/**
+ * HiPanel Base Module
+ * All HiPanel Modules must inherit from here
+ */
+class Module extends \yii\base\Module
+{
+    /// nothing here temporary
+}