+++ common/config/main.php
@@ -1,9 +1,18 @@
 <?php
+/**
+ * @link    http://hiqdev.com/hipanel
+ * @license http://hiqdev.com/hipanel/license
+ * @copyright Copyright (c) 2015 HiQDev
+ */
+
 return [
     'vendorPath' => dirname(dirname(__DIR__)) . '/vendor',
     'components' => [
         'cache' => [
             'class' => 'yii\caching\FileCache',
         ],
+    ],
+    'aliases' => [
+        '@hipanel' => dirname(dirname(__DIR__)),
     ],
 ];