+++ frontend/modules/domain/controllers/HostController.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace frontend\modules\domain\controllers;
-
-class HostController extends \frontend\components\CrudController
-{
-
-}