+++ grid/DataColumn.php
@@ -1,0 +1,16 @@
+<?php
+/**
+ * @link    http://hiqdev.com/hipanel
+ * @license http://hiqdev.com/hipanel/license
+ * @copyright Copyright (c) 2015 HiQDev
+ */
+
+namespace hipanel\grid;
+
+/**
+ * Class DataColumn
+ */
+class DataColumn extends \yii\grid\DataColumn
+{
+    use FeaturedColumnTrait;
+}