+++ widgets/Alert.php
@@ -1,0 +1,90 @@
+<?php
+/**
+ * @link    http://hiqdev.com/hipanel
+ * @license http://hiqdev.com/hipanel/license
+ * @copyright Copyright (c) 2015 HiQDev
+ */
+
+namespace hipanel\widgets;
+
+use raoul2000\widget\pnotify\PNotify;
+
+/**
+ * Alert widget renders a message from session flash. All flash messages are displayed
+ * in the sequence they were assigned using setFlash. You can set message as following:
+ *
+ * ```php
+ * \Yii::$app->getSession()->setFlash('error', 'This is the message');
+ * \Yii::$app->getSession()->setFlash('success', 'This is the message');
+ * \Yii::$app->getSession()->setFlash('info', 'This is the message');
+ * ```
+ *
+ * Multiple messages could be set as follows:
+ *
+ * ```php
+ * \Yii::$app->getSession()->setFlash('error', ['Error 1', 'Error 2']);
+ * ```
+ */
+class Alert extends \yii\bootstrap\Widget
+{
+    /**
+     * @var array the alert types configuration for the flash messages.
+     * This array is setup as $key => $value, where:
+     * - $key is the name of the session flash variable
+     * - $value is the bootstrap alert type (i.e. danger, success, info, warning)
+     */
+    public $alertTypes = [
+        'error'   => 'alert-danger',
+        'danger'  => 'alert-danger',
+        'success' => 'alert-success',
+        'info'    => 'alert-info',
+        'warning' => 'alert-warning'
+    ];
+
+    /**
+     * @var array the options for rendering the close button tag.
+     */
+    public $closeButton = [];
+
+    public function init () {
+        parent::init();
+
+        $session = \Yii::$app->getSession();
+        $flashes = $session->getAllFlashes();
+
+        foreach ($flashes as $type => $data) {
+            if (isset($this->alertTypes[$type])) {
+                $data = (array)$data;
+                foreach ($data as $message) {
+                    $message = $this->normalizeMessage($message);
+
+                    echo PNotify::widget([
+                        'pluginOptions' => array_merge([
+                            'type'    => $type,
+                            'hide'    => true,
+                            'buttons' => [
+                                'sticker' => false
+                            ]
+                        ], $message)
+                    ]);
+                }
+                $session->removeFlash($type);
+            }
+        }
+    }
+
+    /**
+     * @param mixed $message Flash value to be normalized
+     * @return array
+     */
+    public function normalizeMessage ($message) {
+        $res = [];
+        if (is_string($message)) {
+            $res['text'] = $message;
+        } elseif (is_array($message)) {
+            $res = $message;
+        }
+
+        return $res;
+    }
+}