+++ CHANGELOG.md
@@ -1,12 +1,21 @@
 HiPanel changelog
 -----------------
 
-### 0.0.2 under development
+### CURRENT TODO
 
 - moving to plugin modules
+- theme manager
+- menu manager
+- tickets new look
+- spread combo 2 widget
+- REDO EditableColumn, get rid of kartik GridView
+
+### 0.0.2 under development
+
+- GREAT renaming to hipanel
 
 ### 0.0.1-alpha 2015-04-19
 
-- first shown to somebody else :)
+- first shown to someone else :)
 - inited